Fixed admin and user

+++ src/main/java/com/project/parkingmngmtsystem/transaction/Transaction.java
@@ -17,7 +17,7 @@ public class Transaction {
 
     private Long userId;
 
-    private Long spaceId;
+    private Integer spaceId;
     private String carNumber;
 
     private LocalDateTime startTime;
