+++ src/main/java/com/project/parkingmngmtsystem/transaction/TransactionRequest.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Data
 public class TransactionRequest {
     private Long userId;
-    private Long spaceId;
+    private Integer spaceId;
     private String carNumber;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
